Add ImGui, support multiple graphs

+++ shaders/vertex.glsl
@@ -1,7 +1,7 @@
 #version 460 core
 
 layout(std430, binding = 0) volatile buffer gridbuffer {
-	double grid[];
+	float grid[];
 };
 
 uniform mat4 vpmat;
